Issues #15 refactor: 리뷰 반영 + TimeCalculator 패키지 위치 변경
