+++ presentation/src/main/java/com/lighthouse/presentation/binding/TextBindingAdapter.kt
@@ -3,16 +3,17 @@ package com.lighthouse.presentation.binding
 import android.widget.TextView
 import androidx.databinding.BindingAdapter
 import com.lighthouse.presentation.R
-import com.lighthouse.presentation.ui.utils.TimeCalculator
-import com.lighthouse.presentation.ui.utils.TimeCalculator.MAX_DAY
-import com.lighthouse.presentation.ui.utils.TimeCalculator.MIN_DAY
-import java.text.SimpleDateFormat
+import com.lighthouse.presentation.util.TimeCalculator
+import com.lighthouse.presentation.util.TimeCalculator.MAX_DAY
+import com.lighthouse.presentation.util.TimeCalculator.MIN_DAY
+import com.lighthouse.presentation.utils.extention.toDate
+import com.lighthouse.presentation.utils.extention.toMonth
+import com.lighthouse.presentation.utils.extention.toYear
 import java.util.Date
-import java.util.Locale
 
 @BindingAdapter("setExportedDate")
 fun setExportedDate(view: TextView, date: Date) {
-    view.text = date.toString("~ yyyy-MM-dd")
+    view.text = view.context.getString(R.string.all_expired_date, date.toYear(), date.toMonth(), date.toDate())
 }
 
 @BindingAdapter("setDday")
@@ -23,9 +24,4 @@ fun setDday(view: TextView, date: Date) {
         dDay < MIN_DAY -> view.context.getString(R.string.all_d_day_expired)
         else -> view.context.getString(R.string.all_d_day_more_than_year)
     }
-}
-
-private fun Date.toString(format: String): String {
-    val dateFormatter = SimpleDateFormat(format, Locale.getDefault())
-    return dateFormatter.format(this)
 }
