+++ presentation/src/main/java/com/lighthouse/presentation/util/TimeCalculator.kt
@@ -1,4 +1,4 @@
-package com.lighthouse.presentation.ui.utils
+package com.lighthouse.presentation.util
 
 object TimeCalculator {
