+++ presentation/src/main/java/com/lighthouse/presentation/utils/extention/KotlinExtention.kt
@@ -2,8 +2,28 @@ package com.lighthouse.presentation.utils.extention
 
 import android.content.res.Resources
 import android.util.TypedValue
+import java.util.Calendar
+import java.util.Date
 
 val Int.dp
     get() = Resources.getSystem().displayMetrics?.let { dm ->
         TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, this.toFloat(), dm)
     } ?: 0f
+
+fun Date.toYear(): Int {
+    val calendar = Calendar.getInstance()
+    calendar.time = this
+    return calendar.get(Calendar.YEAR)
+}
+
+fun Date.toMonth(): Int {
+    val calendar = Calendar.getInstance()
+    calendar.time = this
+    return calendar.get(Calendar.MONTH)
+}
+
+fun Date.toDate(): Int {
+    val calendar = Calendar.getInstance()
+    calendar.time = this
+    return calendar.get(Calendar.DATE)
+}
